feat: mechanism for empty_value in goalsDB

+++ src/js/goals/db/goalsConnector.js
@@ -2,24 +2,21 @@ import {addDoc, collection, deleteDoc, doc, getDoc, getDocs} from "firebase/fire
 import {db} from "../../firebase.js";
 import {getCurrentUser} from "../../database.js";
 
+export const EMPTY_VALUE = "NONE";
+
+function hasInvalid(...args) {
+    return args.some(v => v === undefined || v === null || v === "" || v === EMPTY_VALUE);
+}
+
 export async function isAuthenticatedUser() {
     const user = await getCurrentUser();
     if (!user) console.error("Utente non autenticato");
     return user
 }
 
-export async function createDocument(collectionName, data) {
-    try {
-        const docRef =
-            await addDoc(collection(db, collectionName), data);
-        return docRef.id;
-    } catch (e) {
-        console.error("Error adding document into " + collectionName + ": ", e);
-        return null;
-    }
-}
-
 export async function clearDocuments(collectionName) {
+    if (hasInvalid(collectionName)) return;
+
     try {
         const querySnapshot = await getDocs(collection(db, collectionName));
         const deletions = querySnapshot.docs.map((doc) => deleteDoc(doc.ref));
@@ -30,7 +27,20 @@ export async function clearDocuments(collectionName) {
     }
 }
 
+export async function createDocument(collectionName, data) {
+    if (hasInvalid(collectionName, data)) return EMPTY_VALUE;
+    try {
+        const docRef =
+            await addDoc(collection(db, collectionName), data);
+        return docRef.id;
+    } catch (e) {
+        console.error("Error adding document into " + collectionName + ": ", e);
+        return EMPTY_VALUE;
+    }
+}
+
 export async function documents(collectionName) {
+    if (hasInvalid(collectionName)) return [];
     try {
         const querySnapshot = await getDocs(collection(db, collectionName));
         const documents = [];
@@ -45,20 +55,22 @@ export async function documents(collectionName) {
 }
 
 export async function buildDocumentRef(collectionName, id) {
+    if (hasInvalid(collectionName, id)) return EMPTY_VALUE;
     try {
         return (await doc(db, collectionName, id));
     } catch (e) {
         console.error("Error building document ref for " + collectionName + " with id " + id + ": ", e);
-        return null;
+        return EMPTY_VALUE;
     }
 }
 
 export async function loadDocumentRef(documentRef) {
+    if (hasInvalid(documentRef)) return EMPTY_VALUE;
     try {
         return (await getDoc(documentRef));
     } catch (e) {
         console.error("Error loading document " + documentRef.id ?? "NONE" + "ref: ", e);
-        return null;
+        return EMPTY_VALUE;
     }
 }
 
@@ -67,6 +79,7 @@ export async function documentsFiltered(collectionName, filterFn) {
 }
 
 export async function documentFromId(collectionName, id) {
-    return (await documentsFiltered(collectionName, document => document.id === id))[0];
+    const firstItem = (await documentsFiltered(collectionName, document => document.id === id))[0];
+    return firstItem ? firstItem : EMPTY_VALUE;
 }
